+++ CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 
 ### Added
+
+* Adds parameter `scale_down_period` to the "buildkite_elastic_stack" module
+
 ### Changed
 ### Deprecated
 ### Removed